Fix login metadata export
